--- catalog/language/en-gb/payment/payfast.php
@@ -1,19 +1,0 @@
-<?php
-
-/**
- * Copyright (c) 2023 PayFast (Pty) Ltd
- * You (being anyone who is not PayFast (Pty) Ltd) may download and use this plugin / code in your own website in
- * conjunction with a registered and active PayFast account. If your PayFast account is terminated for any reason,
- * you may not use this plugin / code or part thereof. Except as expressly indicated in this licence, you may not use,
- * copy, modify or distribute this plugin / code or part thereof in any way.
- */
-
-// Text
-$_['text_title']    = "PayFast";
-$_['text_pay_method'] = "Pay now using PayFast";
-$_['text_reason']   = 'REASON';
-$_['text_sandbox']  = 'Warning: PayFast is set to work in \'Sandbox Mode\', your account will not be charged. ';
-$_['text_total']    = 'Shipping, Handling, Discounts & Taxes';
-$_['text_sale_description'] = 'Shipping, Handling, Discounts and Taxes Included';
-$_['text_recurring']          = '%s every %s %s';
-$_['text_recurring_item']     = 'Recurring Item';
