--- catalog/view/template/payment/payfast.twig
@@ -1,42 +1,0 @@
-{# Copyright (c) 2023 PayFast (Pty) Ltd #}
-{# You (being anyone who is not PayFast (Pty) Ltd) may download and use this plugin / code in your own website in conjunction with a registered and active PayFast account. If your PayFast account is terminated for any reason, you may not use this plugin / code or part thereof. #}
-{# Except as expressly indicated in this licence, you may not use, copy, modify or distribute this plugin / code or part thereof in any way. #}
-
-{% if sandbox == '1' %}
-  <div class="alert alert-danger alert-dismissible"><em class="fa fa-exclamation-circle"></em> {{ text_sandbox }}</div>
-{% endif %}
-<form action="{{ action }}" method="post">
-  <input type="hidden" name="merchant_id" value="{{ merchant_id }}">
-  <input type="hidden" name="merchant_key" value="{{ merchant_key }}">
-  <input type="hidden" name="return_url" value="{{ return_url }}">
-  <input type="hidden" name="cancel_url" value="{{ cancel_url }}">
-  <input type="hidden" name="notify_url" value="{{ notify_url }}">
-  <input type="hidden" name="name_first" value="{{ name_first }}">
-  <input type="hidden" name="name_last" value="{{ name_last  }}">
-  <input type="hidden" name="email_address" value="{{ email_address }}">
-  <input type="hidden" name="m_payment_id" value="{{ m_payment_id }}">
-  <input type="hidden" name="amount" value="{{ amount }}">
-  <input type="hidden" name="item_name" value="{{ item_name }}">
-  <input type="hidden" name="item_description" value="{{ item_description }}">
-  <input type="hidden" name="custom_str1" value="{{ custom_str1 }}">
-
-  {% if recurring %}
-  <input type="hidden" name="custom_str2" value="{{ custom_str2 }}">
-  <input type="hidden" name="custom_str3" value="{{ custom_str3 }}">
-  <input type="hidden" name="custom_str4" value="{{ custom_str4 }}">
-  <input type="hidden" name="custom_str5" value="{{ custom_str5 }}">
-  <input type="hidden" name="subscription_type" value="1">
-  <input type="hidden" name="billing_date" value="{{ billing_date }}">
-  <input type="hidden" name="recurring_amount" value="{{ recurring_amount }}">
-  <input type="hidden" name="frequency" value="{{ frequency }}">
-  <input type="hidden" name="cycles" value="{{ cycles }}">
-  {% endif %}
-
-  <input type="hidden" name="signature" value="{{ signature }}">
-
-  <div class="buttons">
-    <div class="pull-right">
-      <input type="submit" value="{{ button_confirm }}" class="btn btn-primary">
-    </div>
-  </div>
-</form>
