+++ payfast/catalog/controller/payment/composer.json
@@ -1,0 +1,5 @@
+{
+    "require": {
+        "payfast/payfast-common": "^1.0"
+    }
+}
