+++ payfast/install.json
@@ -1,0 +1,8 @@
+{
+	"name": "Payfast",
+    "code": "payfast",
+    "license": "GPL",
+	"version": "1.0.1",
+	"author": "Payfast",
+	"link": "https://payfast.io"
+}
